braille/requirements.txt
pypdf>=4.0.0
EbookLib>=0.18
beautifulsoup4>=4.12.0
python-docx>=1.1.0
pyserial>=3.5
